requirements.txt
# Core ML/DL Frameworks
torch>=2.0.0
torchvision>=0.15.0
tensorflow>=2.13.0
numpy>=1.24.0
scipy>=1.10.0

# Audio Processing
librosa>=0.10.0
soundfile>=0.12.0
audioread>=3.0.0

# Image Processing
Pillow>=10.0.0
opencv-python>=4.8.0
scikit-image>=0.21.0

# XAI Libraries
lime>=0.2.0.1
shap>=0.42.0
captum>=0.6.0  # For Grad-CAM and other attribution methods

# Web Interface (Chainlit)
chainlit>=0.7.0
python-multipart>=0.0.6

# Alternative: Flask (if switching)
# flask>=2.3.0
# flask-cors>=4.0.0

# Data Handling
pandas>=2.0.0
matplotlib>=3.7.0
seaborn>=0.12.0
plotly>=5.15.0

# Utilities
# tqdm>=4.65.0
# python-dotenv>=1.0.0
# pyyaml>=6.0

# # Model Management
# joblib>=1.3.0

# # Testing
# pytest>=7.4.0
# pytest-cov>=4.1.0

# Jupyter (for Colab notebooks)
# jupyter>=1.0.0
# ipywidgets>=8.0.0

# # File handling
# python-magic>=0.4.27  # File type detection

# # Progress bars for web
# stqdm>=0.0.5  # Streamlit-style progress bars

# Optional: for advanced visualizations
# gradio>=3.40.0  # Can be used as alternative interface
# streamlit>=1.25.0  # Can be used as alternative interface
